Moved __tokens file to io.xml package
