+++ src/pysdmx/io/xml/sdmx21/reader/error.py
@@ -2,7 +2,7 @@
 
 from pysdmx.errors import Invalid
 from pysdmx.io.xml.__parse_xml import parse_xml
-from pysdmx.io.xml.sdmx21.__tokens import (
+from pysdmx.io.xml.__tokens import (
     ERROR,
     ERROR_CODE,
     ERROR_MESSAGE,
