+++ src/pysdmx/io/xml/sdmx21/reader/generic.py
@@ -11,7 +11,7 @@ from pysdmx.io.xml.__data_aux import (
     get_data_objects,
 )
 from pysdmx.io.xml.__parse_xml import parse_xml
-from pysdmx.io.xml.sdmx21.__tokens import (
+from pysdmx.io.xml.__tokens import (
     ATTRIBUTES,
     GENERIC,
     ID,
