+++ src/pysdmx/io/xml/sdmx21/reader/submission.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, Sequence
 
 from pysdmx.errors import Invalid
 from pysdmx.io.xml.__parse_xml import parse_xml
-from pysdmx.io.xml.sdmx21.__tokens import (
+from pysdmx.io.xml.__tokens import (
     ACTION,
     MAINTAINABLE_OBJECT,
     REG_INTERFACE,
