+++ src/pysdmx/io/xml/sdmx30/writer/structure_specific.py
@@ -17,7 +17,7 @@ from pysdmx.io.xml.__write_data_aux import (
 from pysdmx.io.xml.__write_structure_specific_aux import (
     __write_data_structure_specific,
 )
-from pysdmx.io.xml.sdmx21.__tokens import (
+from pysdmx.io.xml.__tokens import (
     DSD_LOW,
     PROV_AGREEMENT,
     REGISTRY_LOW,
